ref: search_bar component renamed to searchbar
